refactor(cli): move the process group into its own module
Fourth step of splitting `cli/main.py`. `process` moves to
`geoparquet_io/cli/commands/process.py` as one tree: the `overview` command,
the nested `aggregate` group, and its three subcommands (`a5`, `h3`, `admin`).
A partial move would break dispatch, since `@process.command` and
`@process_aggregate.command` both need their group object in scope.

`_aggregate_error`, which turns a DuckDB binder error from a bad `--where` into
a message naming the option (gpio #612), is used by the three aggregate
subcommands alone, so it travels with the group rather than going to
`_shared.py`.

The moved lines are byte-identical apart from the group's decorator line
(`@cli.group()` -> `@click.group()`, which would otherwise need `cli.main` and
make the dependency cyclic). `main.py` imports the group and attaches it with
`cli.add_command(process)` in the existing registration block, so
`from geoparquet_io.cli.main import process` still resolves; its now-unused
`duckdb`, `core.exceptions` and `core.process.*` imports are dropped.

Also updated: `test_where_error_message_names_the_where_option` imported
`_aggregate_error` from `cli.main`, and now imports it from
`cli.commands.process`. (`test_write_memory_forwarding` and
`test_cli_api_call_parity_scaffold`, the two files #914 parameterised by owning
module, list no `process` case, so neither needed a change.)

`tests/data/cli_surface.json` is unchanged, and `gpio process --help`,
`gpio process overview --help`, `gpio process aggregate --help` and the three
`gpio process aggregate <scheme> --help` pages render identically before and
after. `cli/main.py` is down from 7,349 to 6,595 lines across the two moves.
